HEEDLS-810 - move DisplayStringHelper back to web project
