+++ DigitalLearningSolutions.Data/Helpers/PrnStringHelper.cs
@@ -1,0 +1,12 @@
+﻿namespace DigitalLearningSolutions.Data.Helpers
+{
+    public static class PrnStringHelper
+    {
+        public static string GetPrnDisplayString(bool hasBeenPromptedForPrn, string? professionalRegistrationNumber)
+        {
+            return hasBeenPromptedForPrn
+                ? professionalRegistrationNumber ?? "Not professionally registered"
+                : "Not yet provided";
+        }
+    }
+}
