+++ DigitalLearningSolutions.Web.Tests/Helpers/DisplayStringHelperTests.cs
@@ -1,7 +1,7 @@
-﻿namespace DigitalLearningSolutions.Data.Tests.Helpers
+﻿namespace DigitalLearningSolutions.Web.Tests.Helpers
 {
     using System;
-    using DigitalLearningSolutions.Data.Helpers;
+    using DigitalLearningSolutions.Web.Helpers;
     using FluentAssertions;
     using NUnit.Framework;
 
@@ -281,40 +281,6 @@
 
             // Then
             result.Should().Be(expectedOutput);
-        }
-
-        [Test]
-        public void GetPrnDisplayString_returns_the_prn_when_the_delegate_has_been_prompted_and_has_provided_a_prn()
-        {
-            // Given
-            const string? prn = "12345";
-
-            // When
-            var result = DisplayStringHelper.GetPrnDisplayString(true, prn);
-
-            // Then
-            result.Should().Be(prn);
-        }
-
-        [Test]
-        public void
-            GetPrnDisplayString_returns_Not_professionally_registered_when_the_delegate_has_been_prompted_and_has_not_provided_a_prn()
-        {
-            // When
-            var result = DisplayStringHelper.GetPrnDisplayString(true, null);
-
-            // Then
-            result.Should().Be("Not professionally registered");
-        }
-
-        [Test]
-        public void GetPrnDisplayString_returns_Not_yet_provided_when_the_delegate_has_not_been_prompted_for_a_prn()
-        {
-            // When
-            var result = DisplayStringHelper.GetPrnDisplayString(false, null);
-
-            // Then
-            result.Should().Be("Not yet provided");
         }
 
         [Test]
