+++ DigitalLearningSolutions.Web/ViewModels/Common/Faqs/SearchableFaq.cs
@@ -1,9 +1,9 @@
 namespace DigitalLearningSolutions.Web.ViewModels.Common.Faqs
 {
     using System;
-    using DigitalLearningSolutions.Data.Helpers;
     using DigitalLearningSolutions.Data.Models.SearchSortFilterPaginate;
     using DigitalLearningSolutions.Data.Models.Support;
+    using DigitalLearningSolutions.Web.Helpers;
 
     public class SearchableFaq : BaseSearchableItem
     {
