+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Centre/Administrator/DeactivateAdminViewModel.cs
@@ -1,8 +1,8 @@
 ﻿namespace DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Centre.Administrator
 {
-    using DigitalLearningSolutions.Data.Helpers;
     using DigitalLearningSolutions.Data.Models.User;
     using DigitalLearningSolutions.Web.Attributes;
+    using DigitalLearningSolutions.Web.Helpers;
 
     public class DeactivateAdminViewModel
     {
