+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/CourseSetup/CourseDetails/LearningPathwayDefaultsViewModel.cs
@@ -1,7 +1,7 @@
 ﻿namespace DigitalLearningSolutions.Web.ViewModels.TrackingSystem.CourseSetup.CourseDetails
 {
-    using DigitalLearningSolutions.Data.Helpers;
     using DigitalLearningSolutions.Data.Models.Courses;
+    using DigitalLearningSolutions.Web.Helpers;
 
     public class LearningPathwayDefaultsViewModel
     {
