+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Delegates/CourseDelegates/DelegateProgressViewModel.cs
@@ -4,6 +4,7 @@
     using System.Linq;
     using DigitalLearningSolutions.Data.Helpers;
     using DigitalLearningSolutions.Data.Models.Courses;
+    using DigitalLearningSolutions.Web.Helpers;
     using DigitalLearningSolutions.Web.Models.Enums;
     using DigitalLearningSolutions.Web.ViewModels.Common;
     using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.ViewDelegate;
@@ -19,7 +20,7 @@
         {
             AccessedVia = accessedVia;
             IsCourseActive = details.DelegateCourseInfo.IsCourseActive;
-            ProfessionalRegistrationNumber = DisplayStringHelper.GetPrnDisplayString(
+            ProfessionalRegistrationNumber = PrnStringHelper.GetPrnDisplayString(
                 details.DelegateCourseInfo.HasBeenPromptedForPrn,
                 details.DelegateCourseInfo.ProfessionalRegistrationNumber
             );
