+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Delegates/DelegateProgress/DetailedCourseProgressViewModel.cs
@@ -4,6 +4,7 @@
     using System.Linq;
     using DigitalLearningSolutions.Data.Helpers;
     using DigitalLearningSolutions.Data.Models;
+    using DigitalLearningSolutions.Web.Helpers;
     using DigitalLearningSolutions.Web.Models.Enums;
     using DateHelper = Helpers.DateHelper;
 
@@ -26,7 +27,7 @@
             );
             DelegateEmail = progress.DelegateEmail;
             DelegateNumber = progress.DelegateNumber;
-            ProfessionalRegistrationNumber = DisplayStringHelper.GetPrnDisplayString(
+            ProfessionalRegistrationNumber = PrnStringHelper.GetPrnDisplayString(
                 progress.HasBeenPromptedForPrn,
                 progress.ProfessionalRegistrationNumber
             );
