+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Delegates/GroupDelegates/ConfirmDelegateAddedViewModel.cs
@@ -1,7 +1,7 @@
 ﻿namespace DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.GroupDelegates
 {
-    using DigitalLearningSolutions.Data.Helpers;
     using DigitalLearningSolutions.Data.Models.User;
+    using DigitalLearningSolutions.Web.Helpers;
 
     public class ConfirmDelegateAddedViewModel
     {
