+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Delegates/GroupDelegates/GroupDelegateViewModel.cs
@@ -2,6 +2,7 @@
 {
     using DigitalLearningSolutions.Data.Helpers;
     using DigitalLearningSolutions.Data.Models.DelegateGroups;
+    using DigitalLearningSolutions.Web.Helpers;
 
     public class GroupDelegateViewModel
     {
@@ -14,7 +15,7 @@
             Name = DisplayStringHelper.GetNonSortableFullNameForDisplayOnly(groupDelegate.FirstName, groupDelegate.LastName);
             EmailAddress = groupDelegate.EmailAddress;
             CandidateNumber = groupDelegate.CandidateNumber;
-            ProfessionalRegistrationNumber = DisplayStringHelper.GetPrnDisplayString(
+            ProfessionalRegistrationNumber = PrnStringHelper.GetPrnDisplayString(
                 groupDelegate.HasBeenPromptedForPrn,
                 groupDelegate.ProfessionalRegistrationNumber
             );
