+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Delegates/GroupDelegates/RemoveGroupDelegateViewModel.cs
@@ -1,8 +1,8 @@
 ﻿namespace DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.GroupDelegates
 {
-    using DigitalLearningSolutions.Data.Helpers;
     using DigitalLearningSolutions.Data.Models.DelegateGroups;
     using DigitalLearningSolutions.Web.Attributes;
+    using DigitalLearningSolutions.Web.Helpers;
 
     public class RemoveGroupDelegateViewModel
     {
