+++ DigitalLearningSolutions.Web/ViewModels/TrackingSystem/Delegates/ViewDelegate/WelcomeEmailSentViewModel.cs
@@ -1,7 +1,7 @@
 ﻿namespace DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.ViewDelegate
 {
-    using DigitalLearningSolutions.Data.Helpers;
     using DigitalLearningSolutions.Data.Models.User;
+    using DigitalLearningSolutions.Web.Helpers;
 
     public class WelcomeEmailSentViewModel
     {
