+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/DelegateGroups/ConfirmDeleteGroup.cshtml
@@ -1,5 +1,5 @@
 ﻿@inject IConfiguration Configuration;
-@using DigitalLearningSolutions.Data.Helpers
+@using DigitalLearningSolutions.Web.Helpers
 @using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.DelegateGroups
 @using Microsoft.Extensions.Configuration
 @model ConfirmDeleteGroupViewModel
