+++ DigitalLearningSolutions.Web/Views/TrackingSystem/Delegates/GroupCourses/AddCourseToGroupConfirmation.cshtml
@@ -1,4 +1,4 @@
-﻿@using DigitalLearningSolutions.Data.Helpers
+﻿@using DigitalLearningSolutions.Web.Helpers
 @using DigitalLearningSolutions.Web.ViewModels.TrackingSystem.Delegates.GroupCourses
 @model AddCourseConfirmationViewModel
